fix(signals): move isTestFile/isCodeFile out of local-branch.ts to unbreak ui:typecheck
path-matchers.ts importing isCodeFile/isTestFile from local-branch.ts pulled
the whole review-scoring/Gittensor-API subsystem into any caller's import
graph. focus-manifest.ts (which PR #3690 made import classifyChangedFile
from path-matchers.ts) is reachable from the UI via a long-standing
cross-boundary import, so ui:typecheck started failing repo-wide with
Env/D1Database ambient-type errors it has no way to resolve.

Both functions were only ever thin, self-contained path-matching helpers
incidentally defined in local-branch.ts. Move them into path-matchers.ts
(zero remaining dependency on local-branch.ts), with local-branch.ts
importing them back and re-exporting for its existing importers. Add a
structural test asserting path-matchers.ts never imports from local-branch.ts
again.
