+++ src/signals/path-matchers.ts
@@ -1,10 +1,34 @@
-import { isCodeFile, isTestFile } from "./local-branch";
 import { isTestPath } from "./test-evidence";
 
 // Pure, deterministic path matchers for slop classification (#561). Siblings to `isTestFile` /
 // `isTestPath`: they identify changed files that are NOT genuine hand-authored effort — machine-
 // generated output, vendored/imported third-party code, minified bundles, dependency lockfiles, and
 // docs — so slop signals can tell a padded diff from real work. Path-only and side-effect-free.
+//
+// MUST NOT import from local-branch.ts (#3690-followup): this file is reachable from
+// apps/gittensory-ui/src/lib/registration-workspace.ts via focus-manifest.ts, and local-branch.ts pulls in
+// the whole review-scoring/Gittensor-API subsystem, which breaks `ui:typecheck` under the UI's tsconfig (no
+// Workers ambient types there). local-branch.ts imports isTestFile/isCodeFile back FROM here instead.
+
+/** Keep local scoring aligned with slop/test-evidence matchers (#561 / #1046). */
+export function isTestFile(file: string): boolean {
+  return isTestPath(file);
+}
+
+/** cs/swift/groovy/php plus C/C++/Objective-C round out the native/JVM/.NET/Swift/PHP set: isTestPath already
+ *  recognizes their `SomethingTest(s)`/`Spec` test files, so their source must count as code too —
+ *  otherwise a C#/Swift/Groovy/PHP/native source file is neither test nor code in the local scorer.
+ *  vue/svelte/astro align with review/rag.ts CODE_EXT_RE, review/visual/paths.ts, and rules/advisory.ts
+ *  isCodePath so every classifier agrees. cc/hpp round out the C++ set alongside cpp/c/h (rag.ts already
+ *  indexes all four). dart aligns with rag.ts and test-evidence's *_test.dart convention (hand-authored
+ *  .dart is source; generated .g.dart/.freezed.dart stay non-code via isGeneratedFile). */
+export function isCodeFile(file: string): boolean {
+  return (
+    /\.(ts|tsx|mts|cts|js|jsx|mjs|cjs|py|rb|rs|kt|scala|java|go|sql|cs|swift|groovy|php|cpp|cc|c|h|hpp|m|vue|svelte|astro|dart)$/i.test(
+      file,
+    ) && !isTestFile(file)
+  );
+}
 
 function normalize(path: string): string {
   return String(path ?? "")
